move fakeclock to its own file
